Refactor matches list into shared component, reuse in admin results
Extract ScheduleRow and MatchesList into components/matches/ so the admin
results page can use the same match cards (level badge, referee display,
state-aware score colours) and filters (level + team) as the public dashboard
matches tab. Clicking a card in admin opens the match modal; in the dashboard
it still does nothing.
