+++ frontend/src/pages/tournaments/[id]/dashboard/matches.tsx
@@ -1,19 +1,16 @@
-import { Center, Group, Text } from '@mantine/core';
-import { AiOutlineHourglass } from '@react-icons/all-files/ai/AiOutlineHourglass';
+import { Center, Group } from '@mantine/core';
 import { parseAsInteger, useQueryState } from 'nuqs';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { DashboardFooter } from '@components/dashboard/footer';
 import { DoubleHeader, getTournamentHeadTitle } from '@components/dashboard/layout';
-import { NoContent } from '@components/no_content/empty_table_info';
-import { compareDateTime, formatTime } from '@components/utils/datetime';
+import { MatchesList } from '@components/matches/matches_list';
 import { matchHasTeam } from '@components/utils/match';
 import { responseIsValid, setTitle } from '@components/utils/util';
 import { getStagesLive } from '@services/adapter';
 import { getTournamentResponseByEndpointName } from '@services/dashboard';
 import { getMatchLookup, getStageItemLookup } from '@services/lookups';
-import { ScheduleRow } from './index';
 
 export default function DashboardMatchesPage() {
   const { t } = useTranslation();
@@ -34,70 +31,29 @@ export default function DashboardMatchesPage() {
 
   const stageItemsLookup = getStageItemLookup(swrStagesResponse);
   const matchesLookup = getMatchLookup(swrStagesResponse);
-  const sortedMatches = Object.values(matchesLookup)
-    .filter((item: any) => item.match.start_time != null)
-    .filter((item: any) => levelId == null || item.stage.level_id === levelId)
-    .filter(
-      (item: any) =>
-        teamId == null || matchHasTeam(item.match, teamId, tournamentDataFull.referees_enabled)
-    )
-    .sort(
-      (m1: any, m2: any) =>
-        compareDateTime(m1.match.start_time, m2.match.start_time) ||
-        (m1.match.court?.name || '').localeCompare(m2.match.court?.name || '') ||
-        m1.match.id - m2.match.id
-    );
-
-  const rows: React.JSX.Element[] = [];
-  for (let c = 0; c < sortedMatches.length; c += 1) {
-    const data: any = sortedMatches[c];
-    const startTime = formatTime(data.match.start_time);
-
-    if (
-      c < 1 ||
-      startTime !==
-        formatTime(
-          // sortedMatches only includes matches with a start time (see filter above)
-          sortedMatches[c - 1].match.start_time as string
+  const filteredMatchesLookup =
+    levelId != null || teamId != null
+      ? Object.fromEntries(
+          Object.entries(matchesLookup).filter(
+            ([, entry]: any) =>
+              (levelId == null || entry.stage.level_id === levelId) &&
+              (teamId == null ||
+                matchHasTeam(entry.match, teamId, tournamentDataFull.referees_enabled))
+          )
         )
-    ) {
-      rows.push(
-        <Center mt="md" key={`time-${c}`}>
-          <Text size="xl" fw={800}>
-            {startTime}
-          </Text>
-        </Center>
-      );
-    }
-
-    rows.push(
-      <ScheduleRow
-        key={data.match.id}
-        data={data}
-        stageItemsLookup={stageItemsLookup}
-        matchesLookup={matchesLookup}
-        levels={tournamentDataFull.levels}
-        refereesEnabled={tournamentDataFull.referees_enabled}
-      />
-    );
-  }
+      : matchesLookup;
 
   return (
     <>
       <DoubleHeader tournamentData={tournamentDataFull} />
       <Center>
         <Group style={{ maxWidth: '48rem', width: '100%' }} px="1rem">
-          <div style={{ width: '100%' }}>
-            {rows.length > 0 ? (
-              rows
-            ) : (
-              <NoContent
-                title={t('no_matches_title')}
-                description=""
-                icon={<AiOutlineHourglass />}
-              />
-            )}
-          </div>
+          <MatchesList
+            matchesLookup={filteredMatchesLookup}
+            stageItemsLookup={stageItemsLookup}
+            levels={tournamentDataFull.levels}
+            refereesEnabled={tournamentDataFull.referees_enabled}
+          />
         </Group>
       </Center>
       <DashboardFooter />
